Fjerner felter som er utgått. Oppdaterer tester og validering. Skiller ut validering av fnr til egen fil.
