+++ src/main/kotlin/no/nav/omsorgspenger/soknad/Søknad.kt
@@ -10,23 +10,14 @@ data class Søknad(
     val nyVersjon: Boolean,
     val søknadId: String = UUID.randomUUID().toString(),
     val språk: String,
-    val arbeidssituasjon: List<Arbeidssituasjon>? = null, //TODO 23.02.2021 - Fjernes når frontend er oppdatert
     val kroniskEllerFunksjonshemming: Boolean,
     val barn: BarnDetaljer,
     val sammeAdresse: Boolean?,
     val relasjonTilBarnet: SøkerBarnRelasjon? = null,
     val legeerklæring: List<URL>,
     val samværsavtale: List<URL>? = null,
-    val medlemskap: Medlemskap? = null, //TODO 23.02.2021 - Fjernes når frontend er oppdatert
     val harForståttRettigheterOgPlikter: Boolean,
     val harBekreftetOpplysninger: Boolean
-)
-
-class Medlemskap( //TODO 23.02.2021 - Fjernes når frontend er oppdatert
-    val harBoddIUtlandetSiste12Mnd: Boolean,
-    val utenlandsoppholdSiste12Mnd: List<Utenlandsopphold> = listOf(),
-    val skalBoIUtlandetNeste12Mnd: Boolean,
-    val utenlandsoppholdNeste12Mnd: List<Utenlandsopphold> = listOf()
 )
 
 data class BarnDetaljer(
@@ -41,13 +32,6 @@ data class BarnDetaljer(
     }
 }
 
-data class Utenlandsopphold( //TODO 23.02.2021 - Fjernes når frontend er oppdatert
-    @JsonFormat(pattern = "yyyy-MM-dd") val fraOgMed: LocalDate,
-    @JsonFormat(pattern = "yyyy-MM-dd") val tilOgMed: LocalDate,
-    val landkode: String,
-    val landnavn: String
-)
-
 enum class SøkerBarnRelasjon() {
     @JsonAlias("mor")
     MOR(),
@@ -57,13 +41,4 @@ enum class SøkerBarnRelasjon() {
     ADOPTIVFORELDER(),
     @JsonAlias("fosterforelder")
     FOSTERFORELDER()
-}
-
-enum class Arbeidssituasjon() { //TODO 23.02.2021 - Fjernes når frontend er oppdatert
-    @JsonAlias("selvstendigNæringsdrivende")
-    SELVSTENDIG_NÆRINGSDRIVENDE,
-    @JsonAlias("arbeidstaker")
-    ARBEIDSTAKER,
-    @JsonAlias("frilanser")
-    FRILANSER
 }
