+++ src/main/kotlin/no/nav/omsorgspenger/soknad/SøknadService.kt
@@ -70,11 +70,9 @@ class SøknadService(
             ),
             legeerklæring = legeerklæring,
             samværsavtale = samværsavtale,
-            medlemskap = søknad.medlemskap,
             relasjonTilBarnet = søknad.relasjonTilBarnet,
             harBekreftetOpplysninger = søknad.harBekreftetOpplysninger,
             harForståttRettigheterOgPlikter = søknad.harForståttRettigheterOgPlikter,
-            arbeidssituasjon = søknad.arbeidssituasjon,
             kroniskEllerFunksjonshemming = søknad.kroniskEllerFunksjonshemming,
             nyVersjon = søknad.nyVersjon,
             sammeAdresse = søknad.sammeAdresse,
