+++ src/main/kotlin/no/nav/omsorgspenger/soknad/SøknadValidator.kt
@@ -4,12 +4,6 @@ import no.nav.helse.dusseldorf.ktor.core.*
 import no.nav.k9.søknad.ytelse.omsorgspenger.utvidetrett.v1.OmsorgspengerKroniskSyktBarn
 import no.nav.omsorgspenger.vedlegg.Vedlegg
 import java.net.URL
-import java.time.format.DateTimeFormatter
-
-private val KUN_SIFFER = Regex("\\d+")
-internal val vekttallProviderFnr1: (Int) -> Int = { arrayOf(3, 7, 6, 1, 8, 9, 4, 5, 2).reversedArray()[it] }
-internal val vekttallProviderFnr2: (Int) -> Int = { arrayOf(5, 4, 3, 2, 7, 6, 5, 4, 3, 2).reversedArray()[it] }
-private val fnrDateFormat = DateTimeFormatter.ofPattern("ddMMyy")
 
 private const val MAX_VEDLEGG_SIZE = 24 * 1024 * 1024 // 3 vedlegg på 8 MB
 
@@ -116,7 +110,6 @@ internal fun Søknad.valider(k9FormatSøknad: no.nav.k9.søknad.Søknad) {
 
     violations.addAll(validerK9Format(k9FormatSøknad))
 
-// Ser om det er noen valideringsfeil
     if (violations.isNotEmpty()) {
         throw Throwblem(ValidationProblemDetails(violations))
     }
@@ -150,27 +143,12 @@ private fun BarnDetaljer.valider(relasjonTilBarnet: String?): MutableSet<Violati
             )
         )
     }
-
-    /* //TODO 25.02.2021 - Sett på validering når fjerning av fødselsdato er prodsatt i frontend
     if (norskIdentifikator.isNullOrBlank() && aktørId.isNullOrBlank()) {
         violations.add(
             Violation(
                 parameterName = "barn",
                 parameterType = ParameterType.ENTITY,
                 reason = "Ikke tillatt med barn som mangler norskIdentifikator og aktørID.",
-                invalidValue = norskIdentifikator
-            )
-        )
-    }
-    */
-
-    //TODO 25.02.2021 - Fjerne denne når fødselsdato er fjernet
-    if (norskIdentifikator.isNullOrBlank() && fødselsdato == null && aktørId.isNullOrBlank()) {
-        violations.add(
-            Violation(
-                parameterName = "barn",
-                parameterType = ParameterType.ENTITY,
-                reason = "Ikke tillatt med barn som mangler både fødselsdato, aktørID og norskIdentifikator.",
                 invalidValue = norskIdentifikator
             )
         )
@@ -212,27 +190,6 @@ private fun BarnDetaljer.valider(relasjonTilBarnet: String?): MutableSet<Violati
     }
 
     return violations
-}
-
-fun String.erKunSiffer() = matches(KUN_SIFFER)
-
-fun String.starterMedFodselsdato(): Boolean {
-    // Sjekker ikke hvilket århundre vi skal tolket yy som, kun at det er en gyldig dato.
-    // F.eks blir 290990 parset til 2090-09-29, selv om 1990-09-29 var ønskelig.
-    // Kunne sett på individsifre (Tre første av personnummer) for å tolke århundre,
-    // men virker unødvendig komplekst og sårbart for ev. endringer i fødselsnummeret.
-    return try {
-        var substring = substring(0, 6)
-        val førsteSiffer = (substring[0]).toString().toInt()
-        if (førsteSiffer in 4..7) {
-            substring = (førsteSiffer - 4).toString() + substring(1, 6)
-        }
-        fnrDateFormat.parse(substring)
-
-        true
-    } catch (cause: Throwable) {
-        false
-    }
 }
 
 private fun String.erBlankEllerLengreEnn(maxLength: Int): Boolean = isBlank() || length > maxLength
@@ -278,53 +235,5 @@ fun List<Vedlegg>.validerTotalStørresle() {
     val totalSize = sumBy { it.content.size }
     if (totalSize > MAX_VEDLEGG_SIZE) {
         throw Throwblem(vedleggTooLargeProblemDetails)
-    }
-}
-
-fun String.erGyldigNorskIdentifikator(): Boolean {
-    if (length != 11 || !erKunSiffer() || !starterMedFodselsdato()) return false
-
-    val forventetKontrollsifferEn = get(9)
-
-    val kalkulertKontrollsifferEn = Mod11.kontrollsiffer(
-        number = substring(0, 9),
-        vekttallProvider = vekttallProviderFnr1
-    )
-
-    if (kalkulertKontrollsifferEn != forventetKontrollsifferEn) return false
-
-    val forventetKontrollsifferTo = get(10)
-
-    val kalkulertKontrollsifferTo = Mod11.kontrollsiffer(
-        number = substring(0, 10),
-        vekttallProvider = vekttallProviderFnr2
-    )
-
-    return kalkulertKontrollsifferTo == forventetKontrollsifferTo
-}
-
-/**
- * https://github.com/navikt/helse-sparkel/blob/2e79217ae00632efdd0d4e68655ada3d7938c4b6/src/main/kotlin/no/nav/helse/ws/organisasjon/Mod11.kt
- * https://www.miles.no/blogg/tema/teknisk/validering-av-norske-data
- */
-internal object Mod11 {
-    private val defaultVekttallProvider: (Int) -> Int = { 2 + it % 6 }
-
-    internal fun kontrollsiffer(
-        number: String,
-        vekttallProvider: (Int) -> Int = defaultVekttallProvider
-    ): Char {
-        return number.reversed().mapIndexed { i, char ->
-            Character.getNumericValue(char) * vekttallProvider(i)
-        }.sum().let(Mod11::kontrollsifferFraSum)
-    }
-
-
-    private fun kontrollsifferFraSum(sum: Int) = sum.rem(11).let { rest ->
-        when (rest) {
-            0 -> '0'
-            1 -> '-'
-            else -> "${11 - rest}"[0]
-        }
     }
 }
